fix: refactor after peer review
end date and start date required on controller
end date should be after start date
rename UsageMetricHandler, following naming pattern
fix specs showing date format
